Fix EKF nav estimation
The dimensionality of EKF has been reduced from
15 to 12, as the accelerations have been
removed and the motion estimation consideres a
stable speed throughout the survey or a sensor
capable of correctly estimating the speeds (e.g.
a DVL).
Additionally, this commit adds CSV outputs for
the sensors in the processed folder and a couple
more of formatting corrections.